# ApiContract/quote-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portfolio Club Assurance Quote API",
    "version": "0.1.0",
    "description": "Contract-first OpenAPI document for the ASP.NET Core backend. The shape is compatible with the built-in OpenAPI document generated by modern ASP.NET Core applications using AddOpenApi and MapOpenApi."
  },
  "servers": [
    {
      "url": "http://localhost:5080",
      "description": "Local Docker Compose backend"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Runtime and database health checks."
    },
    {
      "name": "Lookups",
      "description": "Select-box data used by the React quote flow."
    },
    {
      "name": "Quotes",
      "description": "Quote workflow resources."
    },
    {
      "name": "Vehicles",
      "description": "Vehicle creation, listing, and deletion for an auto quote."
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "operationId": "getHealth",
        "summary": "Get API health.",
        "responses": {
          "200": {
            "description": "The API is running.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/health": {
      "get": {
        "tags": ["Health"],
        "operationId": "getDatabaseHealth",
        "summary": "Get quote database health.",
        "responses": {
          "200": {
            "description": "The database is reachable and the quote schema can be inspected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseHealthResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/quote/lookups/vehicle-description": {
      "get": {
        "tags": ["Lookups"],
        "operationId": "getVehicleDescriptionLookups",
        "summary": "Get all select-box data required by the vehicle description step.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Locale"
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle description lookup groups.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDescriptionLookupsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quote/lookups/vehicle-usage": {
      "get": {
        "tags": ["Lookups"],
        "operationId": "getVehicleUsageLookups",
        "summary": "Get all select-box data required by the vehicle usage step.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Locale"
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle usage lookup groups.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleUsageLookupsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quote/lookups/yes-no-unknown": {
      "get": {
        "tags": ["Lookups"],
        "operationId": "getYesNoUnknownOptions",
        "summary": "Get Oui/Non/Je ne sais pas options.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Locale"
          }
        ],
        "responses": {
          "200": {
            "description": "Active yes/no/unknown options.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupOption"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/quote/lookups/purchase-conditions": {
      "get": {
        "tags": ["Lookups"],
        "operationId": "getPurchaseConditionOptions",
        "summary": "Get vehicle purchase condition options.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Locale"
          }
        ],
        "responses": {
          "200": {
            "description": "Active purchase condition options.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupOption"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/quote/lookups/manufacturers": {
      "get": {
        "tags": ["Lookups"],
        "operationId": "getVehicleManufacturers",
        "summary": "Get active vehicle manufacturers.",
        "responses": {
          "200": {
            "description": "Active vehicle manufacturers.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupOption"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/quote/lookups/manufacturers/{manufacturerCode}/models": {
      "get": {
        "tags": ["Lookups"],
        "operationId": "getVehicleModelsByManufacturer",
        "summary": "Get active vehicle models for one manufacturer and optional model year.",
        "parameters": [
          {
            "name": "manufacturerCode",
            "in": "path",
            "required": true,
            "description": "Manufacturer code/name used by the frontend, for example HYUNDAI.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "modelYear",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1980,
              "maximum": 2100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active vehicle models.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleModelOption"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/quotes": {
      "post": {
        "tags": ["Quotes"],
        "operationId": "createAutoQuote",
        "summary": "Create a new auto quote draft.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The quote draft was created.",
            "headers": {
              "Location": {
                "description": "URL of the created quote.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationProblem"
          }
        }
      }
    },
    "/api/quotes/{quoteId}/vehicles": {
      "get": {
        "tags": ["Vehicles"],
        "operationId": "getQuoteVehicles",
        "summary": "List confirmed vehicles attached to a quote.",
        "parameters": [
          {
            "$ref": "#/components/parameters/QuoteId"
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmed vehicles for the quote.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuoteVehicleResponse"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": ["Vehicles"],
        "operationId": "createQuoteVehicle",
        "summary": "Save a confirmed vehicle and its usage answers in the database.",
        "parameters": [
          {
            "$ref": "#/components/parameters/QuoteId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteVehicleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The vehicle was saved for the quote.",
            "headers": {
              "Location": {
                "description": "URL of the saved quote vehicle.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteVehicleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationProblem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/quotes/{quoteId}/vehicles/{vehicleId}": {
      "get": {
        "tags": ["Vehicles"],
        "operationId": "getQuoteVehicleById",
        "summary": "Get one saved vehicle.",
        "parameters": [
          {
            "$ref": "#/components/parameters/QuoteId"
          },
          {
            "$ref": "#/components/parameters/VehicleId"
          }
        ],
        "responses": {
          "200": {
            "description": "The saved vehicle.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteVehicleResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "delete": {
        "tags": ["Vehicles"],
        "operationId": "deleteQuoteVehicle",
        "summary": "Delete a saved vehicle from a quote.",
        "parameters": [
          {
            "$ref": "#/components/parameters/QuoteId"
          },
          {
            "$ref": "#/components/parameters/VehicleId"
          }
        ],
        "responses": {
          "204": {
            "description": "The vehicle was deleted."
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Locale": {
        "name": "locale",
        "in": "query",
        "required": false,
        "description": "Display language for lookup labels. Defaults to fr.",
        "schema": {
          "$ref": "#/components/schemas/Locale"
        }
      },
      "QuoteId": {
        "name": "quoteId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "VehicleId": {
        "name": "vehicleId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Problem Details error response.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "ValidationProblem": {
        "description": "Validation Problem Details response.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationProblemDetails"
            }
          }
        }
      }
    },
    "schemas": {
      "AnnualDistanceKm": {
        "type": "integer",
        "enum": [5000, 10000, 15000, 20000, 25000]
      },
      "CreateQuoteRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "productType": {
            "$ref": "#/components/schemas/ProductType"
          }
        }
      },
      "CreateQuoteVehicleRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "modelYear",
          "manufacturerCode",
          "vehicleModelCode",
          "purchaseYear",
          "purchaseMonth",
          "isLeasedCode",
          "purchaseConditionCode",
          "trackingSystemCode",
          "intensiveEngravingCode",
          "modifiedAfterManufacturingCode",
          "outsideOfProvinceForPersonalUseCode",
          "currentOdometerKm",
          "annualDistanceKm",
          "driveForBusiness"
        ],
        "properties": {
          "modelYear": {
            "$ref": "#/components/schemas/ModelYear"
          },
          "manufacturerCode": {
            "type": "string",
            "examples": ["HYUNDAI"]
          },
          "vehicleModelCode": {
            "type": "string",
            "examples": ["IONIQ_5_PREFERRED_LONG_RANGE_4"]
          },
          "purchaseYear": {
            "$ref": "#/components/schemas/PurchaseYear"
          },
          "purchaseMonth": {
            "$ref": "#/components/schemas/MonthNumber"
          },
          "isLeasedCode": {
            "$ref": "#/components/schemas/YesNoUnknownCode"
          },
          "purchaseConditionCode": {
            "$ref": "#/components/schemas/PurchaseConditionCode"
          },
          "trackingSystemCode": {
            "$ref": "#/components/schemas/YesNoUnknownCode"
          },
          "intensiveEngravingCode": {
            "$ref": "#/components/schemas/YesNoUnknownCode"
          },
          "modifiedAfterManufacturingCode": {
            "$ref": "#/components/schemas/YesNoUnknownCode"
          },
          "outsideOfProvinceForPersonalUseCode": {
            "$ref": "#/components/schemas/YesNoUnknownCode"
          },
          "currentOdometerKm": {
            "$ref": "#/components/schemas/OdometerKm"
          },
          "annualDistanceKm": {
            "$ref": "#/components/schemas/AnnualDistanceKm"
          },
          "driveForBusiness": {
            "type": "boolean"
          }
        }
      },
      "DatabaseHealthResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "schema", "tableCount"],
        "properties": {
          "status": {
            "type": "string",
            "examples": ["ok"]
          },
          "schema": {
            "type": "string",
            "examples": ["quote"]
          },
          "tableCount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "service", "utc"],
        "properties": {
          "status": {
            "type": "string",
            "examples": ["ok"]
          },
          "service": {
            "type": "string",
            "examples": ["PortfolioClubAssurance.Api"]
          },
          "utc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Locale": {
        "type": "string",
        "enum": ["fr", "en"],
        "default": "fr"
      },
      "LookupOption": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "displayText"],
        "properties": {
          "code": {
            "type": "string"
          },
          "displayText": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer"
          }
        }
      },
      "ModelYear": {
        "type": "integer",
        "minimum": 1980,
        "maximum": 2100
      },
      "MonthNumber": {
        "type": "integer",
        "minimum": 1,
        "maximum": 12
      },
      "OdometerKm": {
        "type": "integer",
        "enum": [10000, 30000, 50000, 75000, 100000]
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      },
      "ProductType": {
        "type": "string",
        "enum": ["Auto"],
        "default": "Auto"
      },
      "PurchaseConditionCode": {
        "type": "string",
        "enum": ["New", "Demo", "Used"]
      },
      "PurchaseYear": {
        "type": "integer",
        "minimum": 1900,
        "maximum": 2100
      },
      "QuoteResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "productType", "status", "createdAtUtc", "updatedAtUtc"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productType": {
            "$ref": "#/components/schemas/ProductType"
          },
          "status": {
            "$ref": "#/components/schemas/QuoteStatus"
          },
          "currentStep": {
            "type": ["string", "null"]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "QuoteStatus": {
        "type": "string",
        "enum": ["Draft", "Submitted", "Quoted", "Abandoned"]
      },
      "QuoteVehicleResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "quoteId",
          "modelYear",
          "manufacturer",
          "vehicleModel",
          "purchaseYear",
          "purchaseMonth",
          "isLeased",
          "purchaseCondition",
          "trackingSystem",
          "intensiveEngraving",
          "modifiedAfterManufacturing",
          "outsideOfProvinceForPersonalUse",
          "currentOdometerKm",
          "annualDistanceKm",
          "driveForBusiness",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "quoteId": {
            "type": "string",
            "format": "uuid"
          },
          "modelYear": {
            "$ref": "#/components/schemas/ModelYear"
          },
          "manufacturer": {
            "$ref": "#/components/schemas/LookupOption"
          },
          "vehicleModel": {
            "$ref": "#/components/schemas/VehicleModelOption"
          },
          "vehicleCode": {
            "type": ["string", "null"]
          },
          "purchaseYear": {
            "$ref": "#/components/schemas/PurchaseYear"
          },
          "purchaseMonth": {
            "$ref": "#/components/schemas/MonthNumber"
          },
          "isLeased": {
            "$ref": "#/components/schemas/LookupOption"
          },
          "purchaseCondition": {
            "$ref": "#/components/schemas/LookupOption"
          },
          "trackingSystem": {
            "$ref": "#/components/schemas/LookupOption"
          },
          "intensiveEngraving": {
            "$ref": "#/components/schemas/LookupOption"
          },
          "modifiedAfterManufacturing": {
            "$ref": "#/components/schemas/LookupOption"
          },
          "outsideOfProvinceForPersonalUse": {
            "$ref": "#/components/schemas/LookupOption"
          },
          "currentOdometerKm": {
            "type": "integer",
            "minimum": 0
          },
          "annualDistanceKm": {
            "type": "integer",
            "minimum": 0
          },
          "driveForBusiness": {
            "type": "boolean"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ValidationProblemDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "type": "object",
            "properties": {
              "errors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      },
      "VehicleDescriptionLookupsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "modelYears",
          "purchaseYears",
          "months",
          "manufacturers",
          "yesNoUnknown",
          "purchaseConditions"
        ],
        "properties": {
          "modelYears": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          },
          "purchaseYears": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          },
          "months": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          },
          "manufacturers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          },
          "yesNoUnknown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          },
          "purchaseConditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          }
        }
      },
      "VehicleModelOption": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "displayText", "modelYear", "manufacturerCode"],
        "properties": {
          "code": {
            "type": "string"
          },
          "displayText": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer"
          },
          "modelYear": {
            "$ref": "#/components/schemas/ModelYear"
          },
          "manufacturerCode": {
            "type": "string"
          },
          "trim": {
            "type": ["string", "null"]
          },
          "vehicleCode": {
            "type": ["string", "null"]
          }
        }
      },
      "VehicleUsageLookupsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "provinceUse",
          "odometer",
          "annualDistance",
          "yesNo"
        ],
        "properties": {
          "provinceUse": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          },
          "odometer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          },
          "annualDistance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          },
          "yesNo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOption"
            }
          }
        }
      },
      "YesNoUnknownCode": {
        "type": "string",
        "enum": ["Yes", "No", "Unknown"]
      }
    }
  }
}